backend/auth-service-flask/app/utils/jwt_utils.py
import jwt
from datetime import datetime, timedelta
from flask import current_app

def generate_token(user):
    """Generate JWT token for user"""
    payload = {
        'userId': user.id,
        'username': user.username,
        'email': user.email,
        'roles': user.roles.split(',') if user.roles else [],
        'exp': datetime.utcnow() + timedelta(hours=current_app.config['JWT_EXPIRATION_HOURS']),
        'iat': datetime.utcnow()
    }
    
    token = jwt.encode(
        payload,
        current_app.config['JWT_SECRET_KEY'],
        algorithm='HS256'
    )
    
    return token

def decode_token(token):
    """Decode and validate JWT token"""
    try:
        payload = jwt.decode(
            token,
            current_app.config['JWT_SECRET_KEY'],
            algorithms=['HS256']
        )
        return payload
    except jwt.ExpiredSignatureError:
        raise ValueError('Token has expired')
    except jwt.InvalidTokenError:
        raise ValueError('Invalid token')
