backend/auth-service-flask/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-migrate==4.0.5
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
marshmallow==3.20.1
flasgger==0.9.7.1
gunicorn==21.2.0
